move parsers to module
