+++ lib/freshbooks_api_client/interface/tasks.ex
@@ -6,8 +6,6 @@ defmodule FreshbooksApiClient.Interface.Tasks do
   It uses a FreshbooksApiClient.Interface
   """
 
-  import SweetXml
-
   use FreshbooksApiClient.Interface,
     schema: FreshbooksApiClient.Schema.Task,
     resources: "tasks",
